+++ src/Controller/FeedGetDataController.php
@@ -6,10 +6,11 @@ namespace App\Controller;
 
 use App\Entity\Tenant\Feed;
 use App\Service\FeedService;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpKernel\Attribute\AsController;
 
 #[AsController]
-final readonly class FeedGetDataController
+final class FeedGetDataController extends AbstractController
 {
     public function __construct(
         private FeedService $feedService