+++ src/State/FeedProvider.php
@@ -6,39 +6,31 @@ namespace App\State;
 
 use ApiPlatform\Doctrine\Orm\Extension\QueryItemExtensionInterface;
 use ApiPlatform\Doctrine\Orm\Util\QueryNameGenerator;
-use ApiPlatform\Metadata\Get;
 use ApiPlatform\Metadata\Operation;
 use ApiPlatform\State\ProviderInterface;
 use App\Entity\Tenant\Feed;
-use App\Entity\Tenant\FeedData;
 use App\Entity\User;
 use App\Repository\FeedRepository;
 use App\Repository\PlaylistSlideRepository;
 use App\Service\FeedService;
 use Doctrine\ORM\NonUniqueResultException;
-use Psr\Log\LoggerInterface;
 use Symfony\Bundle\SecurityBundle\Security;
-use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\Uid\Ulid;
-use Symfony\Contracts\HttpClient\Exception\ClientExceptionInterface;
-use Symfony\Contracts\HttpClient\Exception\RedirectionExceptionInterface;
-use Symfony\Contracts\HttpClient\Exception\ServerExceptionInterface;
 
 /**
- * A FeedData state provider.
+ * A Feed state provider.
  *
  * @see https://api-platform.com/docs/v2.7/core/state-providers/
  *
- * @template T of FeedData
+ * @template T of Feed
  */
-final class FeedDataProvider extends AbstractProvider
+final class FeedProvider extends AbstractProvider
 {
     public function __construct(
         private readonly Security $security,
         private readonly PlaylistSlideRepository $playlistSlideRepository,
         private readonly FeedRepository $feedRepository,
         private readonly FeedService $feedService,
-        private readonly LoggerInterface $logger,
         private readonly iterable $itemExtensions,
         ProviderInterface $collectionProvider
     ) {
@@ -118,23 +110,6 @@ final class FeedDataProvider extends AbstractProvider
             }
         }
 
-        if (!is_null($feed)) {
-            $feedId = $feed->getId();
-            $feedIdReference = null === $feedId ? '0' : $feedId->jsonSerialize();
-            try {
-                return new JsonResponse($this->feedService->getData($feed), \Symfony\Component\HttpFoundation\Response::HTTP_OK);
-            } catch (MissingFeedConfigurationException $e) {
-                $this->logger->error(sprintf('Missing configuration for feed with id "%s" with message "%s"', $feedIdReference, $e->getMessage()));
-            } catch (\JsonException $e) {
-                $this->logger->error(sprintf('JSON decode for feed with id "%s" with error "%s"', $feedIdReference, $e->getMessage()));
-            } catch (ClientExceptionInterface|RedirectionExceptionInterface|ServerExceptionInterface|TransportExceptionInterface $e) {
-                $this->logger->error(sprintf('Communication error "%s"', $e->getMessage()));
-            } catch (\Throwable $e) {
-                $this->logger->error(sprintf('Feed data error. ID: %s, MESSAGE: %s', $feed->getId()->jsonSerialize(), $e->getMessage()));
-            }
-        }
-
-        // Null returned for data provider will result in a 404 response from API platform.
-        return null;
+        return $feed;
     }
 }